vel 14 50 to 60

+++ hypercuber.py
@@ -201,17 +201,21 @@ hcube.load_nhi_rad_857(local=False)
 
 for _v in [14]: # of 21
     print("running velocity {}".format(_v))
-    for _thet in np.arange(10, 20): # of 165
-        time0 = time.time()
-        hcube.tabulate_per_vel_theta(vel_i=_v, theta_i=_thet, verbose=False)
-        time1 = time.time()
+    for _thet in np.arange(50, 60): # of 165
+    
+        if os.path.isfile("temp_hcube_slices/hypercube_nhi_v{}_t{}.npy".format(_v, _thet)):
+            print("v {}, t {} already exists".format(_v, _thet))
+        else:
+            time0 = time.time()
+            hcube.tabulate_per_vel_theta(vel_i=_v, theta_i=_thet, verbose=False)
+            time1 = time.time()
         
-        print("finished with velocity {} of 21, thet {} of 165. Took {} min.".format(_v + 1, _thet + 1, (time1-time0)/60.))
+            print("finished with velocity {} of 21, thet {} of 165. Took {} min.".format(_v + 1, _thet + 1, (time1-time0)/60.))
 
-        np.save("temp_hcube_slices/hypercube_nhi_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_nhi[:, :, _v, _thet])
-        np.save("temp_hcube_slices/hypercube_rad_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_rad[:, :, _v, _thet])
-        np.save("temp_hcube_slices/hypercube_857_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_857[:, :, _v, _thet])
-        np.save("temp_hcube_slices/hypercube_weights_v{}_t{}.npy".format(_v, _thet), hcube.weights_hypercube[:, :, _v, _thet])
+            np.save("temp_hcube_slices/hypercube_nhi_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_nhi[:, :, _v, _thet])
+            np.save("temp_hcube_slices/hypercube_rad_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_rad[:, :, _v, _thet])
+            np.save("temp_hcube_slices/hypercube_857_v{}_t{}.npy".format(_v, _thet), hcube.hypercube_857[:, :, _v, _thet])
+            np.save("temp_hcube_slices/hypercube_weights_v{}_t{}.npy".format(_v, _thet), hcube.weights_hypercube[:, :, _v, _thet])
 
 
 hcube = HyperCube(singlecube=False)
